fix: Store & Order 서비스 분리
